# multiagent/console_ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VF O2O Console API",
    "version": "0.1.0",
    "description": "Hop dong cho frontend Console tai /console. Xac thuc bang cookie phien HttpOnly (same-origin); moi POST phai gui header X-CSRF-Token lay tu GET /auth/me."
  },
  "paths": {
    "/api/console/v1/auth/login": {
      "post": {
        "tags": [
          "console"
        ],
        "summary": "Login",
        "operationId": "login_api_console_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/console/v1/reviews/{public_id}": {
      "get": {
        "tags": [
          "console"
        ],
        "summary": "Get Review",
        "operationId": "get_review_api_console_v1_reviews__public_id__get",
        "parameters": [
          {
            "name": "public_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Public Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDetailModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentResultModel": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "integer"
              },
              {
                "type": "string"
              },
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "criteria": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Criteria"
          },
          "issues": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Issues"
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "name",
          "score",
          "criteria",
          "issues",
          "evidence"
        ],
        "title": "AgentResultModel"
      },
...
      "CostEstimateModel": {
        "properties": {
          "input_tokens": {
            "type": "integer",
            "title": "Input Tokens"
          },
          "output_tokens": {
            "type": "integer",
            "title": "Output Tokens"
          },
          "estimated_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Usd"
          },
          "pricing_version": {
            "type": "integer",
            "title": "Pricing Version"
          },
          "effective_at": {
            "type": "string",
            "title": "Effective At"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "unknown_models": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unknown Models"
          }
        },
        "type": "object",
        "required": [
          "input_tokens",
          "output_tokens",
          "estimated_usd",
          "pricing_version",
          "effective_at",
          "currency",
          "source",
          "unknown_models"
        ],
        "title": "CostEstimateModel"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "default": ""
          },
          "password": {
            "type": "string",
            "title": "Password",
            "default": ""
          }
        },
        "type": "object",
        "title": "LoginRequest"
      },
      "MeResponse": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "must_change_password": {
            "type": "boolean",
            "title": "Must Change Password"
          },
          "csrf_token": {
            "type": "string",
            "title": "Csrf Token"
          }
        },
        "type": "object",
        "required": [
          "username",
          "role",
          "must_change_password",
          "csrf_token"
        ],
        "title": "MeResponse"
      },
...
      "ReviewDetailModel": {
        "properties": {
          "public_id": {
            "type": "string",
            "title": "Public Id"
          },
          "scored_at": {
            "type": "string",
            "title": "Scored At"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "final_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Score"
          },
          "missing_agents": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Agents"
          },
          "veto_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Veto Reason"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "agents": {
            "items": {
              "$ref": "#/components/schemas/AgentResultModel"
            },
            "type": "array",
            "title": "Agents"
          },
          "config_meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Meta"
          },
          "cost_estimate": {
            "$ref": "#/components/schemas/CostEstimateModel"
          },
          "usage_available": {
            "type": "boolean",
            "title": "Usage Available"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "writeback_status": {
            "type": "string",
            "title": "Writeback Status"
          },
          "writeback_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Writeback Error"
          },
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "site_slug": {
            "type": "string",
            "title": "Site Slug"
          },
          "site_name": {
            "type": "string",
            "title": "Site Name"
          },
          "profile_id": {
            "type": "string",
            "title": "Profile Id"
          },
          "profile_code": {
            "type": "string",
            "title": "Profile Code"
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version"
          },
          "external_content_id": {
            "type": "string",
            "title": "External Content Id"
          },
          "external_revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Revision Id"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "langcode": {
            "type": "string",
            "title": "Langcode"
          },
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id"
          },
          "is_fixture": {
            "type": "boolean",
            "title": "Is Fixture"
          },
          "drupal_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drupal Url"
          }
        },
        "type": "object",
        "required": [
          "public_id",
          "scored_at",
          "duration_ms",
          "decision",
          "final_score",
          "missing_agents",
          "veto_reason",
          "note",
          "agents",
          "config_meta",
          "cost_estimate",
          "usage_available",
          "model",
          "writeback_status",
          "writeback_error",
          "site_id",
          "site_slug",
          "site_name",
          "profile_id",
          "profile_code",
          "policy_version",
          "external_content_id",
          "external_revision_id",
          "content_type",
          "langcode",
          "correlation_id",
          "is_fixture",
          "drupal_url"
        ],
        "title": "ReviewDetailModel"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}